Implement dynamic home screen data fetching and enhance cashflow filtering
- Integrate `GetHomeSummaryUseCase` and `GetRecentCashFlowsUseCase` into `HomeViewModel` to fetch real-time financial data.
- Update `HomeScreen` and `CashFlowScreen` to support filtering by both month and year.
- Refactor `AnalyticNetworkDatasource` and `CashFlowNetworkDatasource` to include `user_id` filtering from the current Supabase session.
- Enhance `DashboardAppBar` to display the wallet balance dynamically and sync month selection across screens.
- Implement currency formatting for income, expenses, and balances using `NumberFormat`.
- Move `YearSelector` to core components for shared use across analytic, home, and cashflow screens.
- Update `HomeScreenState` to include `isLoading`, `year`, and use `Double` for financial values.
- Register new home-related data sources, repositories, and use cases in `AppModule`.

+++ app/core/src/main/java/com/example/core/components/CardBudget.kt
@@ -30,7 +30,7 @@ fun CardBudget(
     category: String = "",
     icon: ImageVector = Icons.Rounded.ShoppingCart,
     info: String? = "",
-    amount: Int = 0,
+    amount: Double = 0.0,
     tint: Color = PrimaryBlue,
 ) {
     val formattedCurrency = NumberFormat.getCurrencyInstance(Locale("id", "ID")).format(amount)
